requirements.txt
pefile==2019.4.18
termcolor==1.1.0
python-magic==0.4.15
ssdeep==3.3
coverage==4.5.4
pytest==5.1.2
coverage==4.5.4
pytest-cov==2.7.1
coveralls